% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psarras_adaptation-package.R
\docType{package}
\name{psarrasadaptation-package}
\alias{psarrasadaptation}
\alias{psarrasadaptation-package}
\title{psarrasadaptation: Adaptation of linear systems benchmark from Psarras et al. 2022. <doi:10.1145/2F3549935>}
\description{
Compare base R, Armadillo and Eigen for solving linear systems of equations. The benchmark is done by using a direct inv(A) * b operation and the solve methods that each library provides for compiler optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11eigen}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11eigen/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
